# api/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: Animus Gateway API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /auth/session:
    get:
      summary: Get current session identity
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/login:
    get:
      summary: Start OIDC login flow
      tags: [Auth]
      parameters:
        - name: return_to
          in: query
          required: false
          schema:
            type: string
          description: Relative path to redirect to after login.
      responses:
        "302":
          description: Redirect to OIDC provider
        "501":
          description: Login flow not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/callback:
    get:
      summary: OIDC callback endpoint
      tags: [Auth]
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect back to UI
        "400":
          description: Invalid callback parameters/state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "501":
          description: Login flow not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/audit/{proxyPath}:
    parameters:
      - name: proxyPath
        in: path
        required: true
        schema:
          type: string
        description: Remaining path to proxy (may include slashes).
    get:
      summary: Proxy GET to Audit
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    post:
      summary: Proxy POST to Audit
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    put:
      summary: Proxy PUT to Audit
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    patch:
      summary: Proxy PATCH to Audit
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
    delete:
      summary: Proxy DELETE to Audit
      tags: [Proxy]
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ProxiedOK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/BadGateway"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ProxiedOK:
      description: Proxied response from upstream service
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Upstream error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    SessionResponse:
      type: object
      additionalProperties: false
      required: [subject, email, roles]
      properties:
        subject:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        request_id:
          type: string